1.9

+++ src/main/java/cn/tedu/jsdvn2203/csmall/server/service/impl/AlbumServiceImpl.java
@@ -1,5 +1,6 @@
 package cn.tedu.jsdvn2203.csmall.server.service.impl;
 
+import cn.tedu.jsdvn2203.csmall.server.exception.ServiceException;
 import cn.tedu.jsdvn2203.csmall.server.mapper.AlbumMapper;
 import cn.tedu.jsdvn2203.csmall.server.pojo.dto.AlbumAddNewDTO;
 import cn.tedu.jsdvn2203.csmall.server.pojo.dto.AlbumDeleteDTO;
@@ -22,6 +23,14 @@ public class AlbumServiceImpl implements IAlbumService {
     AlbumMapper albumMapper;
     @Override
     public void addNew(AlbumAddNewDTO albumAddNewDTO){
+        String name = albumAddNewDTO.getName();
+        int count = albumMapper.countByName(name);
+        if (count > 0){
+            String message = "添加失败相册名称["+name+"]已存在";
+            log.error(message);
+            throw new ServiceException(message);
+        }
+
         //创建相册实体
         Album album = new Album();
         BeanUtils.copyProperties(albumAddNewDTO,album);
@@ -31,6 +40,14 @@ public class AlbumServiceImpl implements IAlbumService {
 
     @Override
     public void delete(AlbumDeleteDTO albumDeleteDTO) {
+        Long id = albumDeleteDTO.getId();
+        int count = albumMapper.countById(id);
+        if (count == 0){
+            String message = "删除失败，相册id["+id+"]不存在";
+            log.error(message);
+            throw new ServiceException(message);
+        }
+
         Album album = new Album();
         BeanUtils.copyProperties(albumDeleteDTO,album);
         int rows = albumMapper.deleteById(album.getId());
